♻️ Move context files in a dedicated folder and collocate their types
