+++ packages/rum-core/src/boot/startRum.ts
@@ -3,10 +3,10 @@ import { startTelemetry, canUseEventBridge, getEventBridge, startFlushFailedSend
 import { createDOMMutationObservable } from '../browser/domMutationObservable'
 import { startPerformanceCollection } from '../browser/performanceCollection'
 import { startRumAssembly } from '../domain/assembly'
-import { startForegroundContexts } from '../domain/foregroundContexts'
-import { startInternalContext } from '../domain/internalContext'
+import { startForegroundContexts } from '../domain/contexts/foregroundContexts'
+import { startInternalContext } from '../domain/contexts/internalContext'
 import { LifeCycle } from '../domain/lifeCycle'
-import { startViewContexts } from '../domain/viewContexts'
+import { startViewContexts } from '../domain/contexts/viewContexts'
 import { startRequestCollection } from '../domain/requestCollection'
 import { startActionCollection } from '../domain/rumEventsCollection/action/actionCollection'
 import { startErrorCollection } from '../domain/rumEventsCollection/error/errorCollection'
@@ -18,7 +18,7 @@ import { startRumSessionManager, startRumSessionManagerStub } from '../domain/ru
 import type { CommonContext } from '../rawRumEvent.types'
 import { startRumBatch } from '../transport/startRumBatch'
 import { startRumEventBridge } from '../transport/startRumEventBridge'
-import { startUrlContexts } from '../domain/urlContexts'
+import { startUrlContexts } from '../domain/contexts/urlContexts'
 import type { LocationChange } from '../browser/locationChangeObservable'
 import { createLocationChangeObservable } from '../browser/locationChangeObservable'
 import type { RumConfiguration } from '../domain/configuration'
