+++ packages/rum-core/src/domain/assembly.ts
@@ -22,14 +22,14 @@ import type {
 } from '../rawRumEvent.types'
 import { RumEventType } from '../rawRumEvent.types'
 import type { RumEvent } from '../rumEvent.types'
-import { getSyntheticsContext } from './syntheticsContext'
-import { getCiTestContext } from './ciTestContext'
+import { getSyntheticsContext } from './contexts/syntheticsContext'
+import { getCiTestContext } from './contexts/ciTestContext'
 import type { LifeCycle } from './lifeCycle'
 import { LifeCycleEventType } from './lifeCycle'
-import type { ViewContexts } from './viewContexts'
+import type { ViewContexts } from './contexts/viewContexts'
 import type { RumSessionManager } from './rumSessionManager'
 import { RumSessionPlan } from './rumSessionManager'
-import type { UrlContexts } from './urlContexts'
+import type { UrlContexts } from './contexts/urlContexts'
 import type { RumConfiguration } from './configuration'
 import type { ActionContexts } from './rumEventsCollection/action/actionCollection'
