+++ packages/rum-core/src/domain/contexts/foregroundContexts.ts
@@ -1,6 +1,6 @@
 import type { RelativeTime, Duration } from '@datadog/browser-core'
 import { addEventListener, DOM_EVENT, elapsed, relativeNow, toServerDuration } from '@datadog/browser-core'
-import type { InForegroundPeriod } from '../rawRumEvent.types'
+import type { InForegroundPeriod } from '../../rawRumEvent.types'
 
 // Arbitrary value to cap number of element mostly for backend & to save bandwidth
 export const MAX_NUMBER_OF_SELECTABLE_FOREGROUND_PERIODS = 500
