+++ packages/rum-core/src/domain/contexts/syntheticsContext.ts
