+++ packages/rum-core/src/domain/contexts/urlContexts.spec.ts
@@ -1,11 +1,11 @@
 import type { RelativeTime } from '@datadog/browser-core'
 import { relativeToClocks } from '@datadog/browser-core'
-import type { TestSetupBuilder } from '../../test/specHelper'
-import { setup } from '../../test/specHelper'
+import type { TestSetupBuilder } from '../../../test/specHelper'
+import { setup } from '../../../test/specHelper'
+import { LifeCycleEventType } from '../lifeCycle'
+import type { ViewCreatedEvent, ViewEndedEvent } from '../rumEventsCollection/view/trackViews'
 import type { UrlContexts } from './urlContexts'
 import { startUrlContexts } from './urlContexts'
-import type { ViewCreatedEvent, ViewEndedEvent } from './rumEventsCollection/view/trackViews'
-import { LifeCycleEventType } from './lifeCycle'
 
 describe('urlContexts', () => {
   let setupBuilder: TestSetupBuilder
