+++ packages/rum-core/src/domain/rumEventsCollection/view/viewCollection.ts
@@ -5,7 +5,7 @@ import type { RawRumViewEvent } from '../../../rawRumEvent.types'
 import { RumEventType } from '../../../rawRumEvent.types'
 import type { LifeCycle, RawRumEventCollectedData } from '../../lifeCycle'
 import { LifeCycleEventType } from '../../lifeCycle'
-import type { ForegroundContexts } from '../../foregroundContexts'
+import type { ForegroundContexts } from '../../contexts/foregroundContexts'
 import type { LocationChange } from '../../../browser/locationChangeObservable'
 import type { RumConfiguration } from '../../configuration'
 import type { ViewEvent, ViewOptions } from './trackViews'
