+++ packages/rum-core/src/index.ts
@@ -18,11 +18,11 @@ export {
   RumViewEventDomainContext,
   RumEventDomainContext,
 } from './domainContext.types'
-export { ViewContext, CommonContext, ReplayStats } from './rawRumEvent.types'
+export { CommonContext, ReplayStats } from './rawRumEvent.types'
 export { startRum } from './boot/startRum'
 export { LifeCycle, LifeCycleEventType } from './domain/lifeCycle'
 export { ViewCreatedEvent } from './domain/rumEventsCollection/view/trackViews'
-export { ViewContexts } from './domain/viewContexts'
+export { ViewContexts, ViewContext } from './domain/contexts/viewContexts'
 export { RumSessionManager, RumSessionPlan } from './domain/rumSessionManager'
 export { getMutationObserverConstructor } from './browser/domMutationObservable'
 export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
