+++ packages/rum-core/src/rawRumEvent.types.ts
@@ -198,6 +198,9 @@ export interface RumContext {
     test_id: string
     result_id: string
   }
+  ci_test?: {
+    test_execution_id: string
+  }
   _dd: {
     format_version: 2
     drift: number
@@ -205,45 +208,6 @@ export interface RumContext {
       plan: RumSessionPlan
     }
     browser_sdk_version?: string
-  }
-  ci_test?: {
-    test_execution_id: string
-  }
-}
-
-export interface ViewContext extends Context {
-  service?: string
-  version?: string
-  view: {
-    id: string
-    name?: string
-  }
-}
-
-export interface ActionContext extends Context {
-  action: {
-    id: string | string[]
-  }
-}
-
-export interface UrlContext extends Context {
-  view: {
-    url: string
-    referrer: string
-  }
-}
-
-export interface InternalContext {
-  application_id: string
-  session_id: string | undefined
-  view?: {
-    id: string
-    url: string
-    referrer: string
-    name?: string
-  }
-  user_action?: {
-    id: string | string[]
   }
 }
