+++ packages/rum-core/test/specHelper.ts
@@ -3,24 +3,24 @@ import { assign, combine, Observable, noop, setCookie, deleteCookie, ONE_MINUTE 
 import type { Clock } from '../../core/test/specHelper'
 import { SPEC_ENDPOINTS, mockClock, buildLocation } from '../../core/test/specHelper'
 import type { RecorderApi } from '../src/boot/rumPublicApi'
-import type { ForegroundContexts } from '../src/domain/foregroundContexts'
+import type { ForegroundContexts } from '../src/domain/contexts/foregroundContexts'
 import type { RawRumEventCollectedData } from '../src/domain/lifeCycle'
 import { LifeCycle, LifeCycleEventType } from '../src/domain/lifeCycle'
-import type { ViewContexts } from '../src/domain/viewContexts'
+import type { ViewContext, ViewContexts } from '../src/domain/contexts/viewContexts'
 import type { ViewEvent, ViewOptions } from '../src/domain/rumEventsCollection/view/trackViews'
 import { trackViews } from '../src/domain/rumEventsCollection/view/trackViews'
 import type { RumSessionManager } from '../src/domain/rumSessionManager'
 import { RumSessionPlan } from '../src/domain/rumSessionManager'
-import type { RawRumEvent, RumContext, ViewContext, UrlContext } from '../src/rawRumEvent.types'
+import type { RawRumEvent, RumContext } from '../src/rawRumEvent.types'
 import type { LocationChange } from '../src/browser/locationChangeObservable'
-import type { UrlContexts } from '../src/domain/urlContexts'
-import type { BrowserWindow } from '../src/domain/syntheticsContext'
+import type { UrlContext, UrlContexts } from '../src/domain/contexts/urlContexts'
+import type { BrowserWindow } from '../src/domain/contexts/syntheticsContext'
 import {
   SYNTHETICS_INJECTS_RUM_COOKIE_NAME,
   SYNTHETICS_RESULT_ID_COOKIE_NAME,
   SYNTHETICS_TEST_ID_COOKIE_NAME,
-} from '../src/domain/syntheticsContext'
-import type { CiTestWindow } from '../src/domain/ciTestContext'
+} from '../src/domain/contexts/syntheticsContext'
+import type { CiTestWindow } from '../src/domain/contexts/ciTestContext'
 import type { RumConfiguration } from '../src/domain/configuration'
 import { validateAndBuildRumConfiguration } from '../src/domain/configuration'
 import type { ActionContexts } from '../src/domain/rumEventsCollection/action/actionCollection'
